requirements.txt
mediapipe
opencv-python
tensorflow
numpy

# For Ai-bot
openai-whisper
transformers
langchain
langchain-community
langchain-huggingface
huggingface_hub
sounddevice
soundfile
numpy
requests
keyboard
python-dotenv
sounddevice
python-dotenv
langchain
langchain-groq
edge-tts
pygame
numpy

# for FastAPI
fastapi
uvicorn
requests
python-multipart

# for DB
Flask==2.0.1
psycopg2==2.9.9
python-dotenv==0.19.0
Werkzeug==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
email-validator==1.1.3
boto3